Move remove_host_context
