+++ frontend/catalyst/utils/jax_extras.py
@@ -636,52 +636,3 @@ def _gather_shape_rule_dynamic(
             )
 
     return _gather_shape_computation(indices, dimension_numbers, slice_sizes)
-
-
-def remove_host_context(jaxpr):
-    """This function will remove the host context.
-
-    It is expected that callers to this function do not use the host context
-    and instead forward the arguments directly to the qnode.
-    This is useful for example in the proof-of-concept for cuda integration.
-
-    Later iterations **might** have a host context. This is something not currently planned.
-
-    The host context is the wrapper function that calls a qnode. When applying a `@qjit`
-    decorator immediately on a qnode like:
-
-    ```python
-    @qjit
-    @qml.qnode(qml.device("lightning.qubit", wires=1))
-    def identity(x):
-      return x
-    ```
-
-    Notice that in the JAXPR (and all subsequent IRs), we have this wrapper context that only
-    performs a function call.
-
-    { lambda ; a:i64[]. let
-      b:i64[] = func[
-        call_jaxpr={ lambda ; c:i64[]. let
-             = qdevice[
-              rtd_kwargs={'shots': 0, 'mcmc': False}
-              rtd_lib=path/to/runtime...
-              rtd_name=LightningSimulator
-            ]
-            d:AbstractQreg() = qalloc 1
-             = qdealloc d
-          in (c,) }
-        fn=<QNode: wires=1, device='lightning.qubit', interface='auto', diff_method='best'>
-      ] a
-    in (b,) }
-    """
-    # Is this always the case, let's try with different parameters.
-    is_one_equation = len(jaxpr.jaxpr.eqns) == 1
-    from catalyst.jax_primitives import func_p
-    from catalyst.utils.exceptions import CompileError
-
-    is_single_equation_call = jaxpr.jaxpr.eqns[0].primitive == func_p
-    is_valid = is_one_equation and is_single_equation_call
-    if is_valid:
-        return jaxpr.jaxpr.eqns[0][3]["call_jaxpr"]
-    raise CompileError("Cannot translate tapes with context.")
